refactor: move remove unsupported data types
